Nav and Footer Update
I started by organizing files in components and pages to better fit out current format. I followed this up by organizing imports at the top of each page with comments so that we all know what is what and keeps things clean! I switched the nav out and have not put information in for it yet (I plan on doing this tonight/tomorrow before class). I updated the Nav again to be a lot more fluid and to soon match the footer.